Allow NVIDIA_VISIBLE_DEVICES=void to always override
This change adjusts the order of operations of the NVIDIA_VISIBLE_DEVICES
envvar values. This is especially important for cases where duplicates
may exist. Previously a value of all would be preferred over other values
meaning that even if a user were to set NVIDIA_VISIBLE_DEVICES=void,
depending on the runtime behaviour the value of NVIDIA_VISIBLE_DEVICES=all
may still take precedence.

With this change, the order of precendece is updated to void > none > all.

Signed-off-by: Evan Lezar <elezar@nvidia.com>