Generate separate specs for coherent and noncoherent devices
With this change the nvidia-ctk cdi generate command generates
CDI specs based on whether a device supports coherent access to
system memory or not. In this case "regular" nvidia.com/gpu CDI
specs are generated for all devices as well as
nvidia.com/gpu.coherent and nvidia.com/gpu.noncoherent for devices
that are either coherent or non-coherent.

Adding the --feature-flag=disable-coherent-annotations command line
argument to the nvidia-ctk cdi generate command will disable this.

The "disable-coherent-annotations" feature flag can also be set in the
nvcdi API in which case the generated CDI device specification will
not include annotations indicating coherence.

Signed-off-by: Evan Lezar <elezar@nvidia.com>