[no-relnote] Add e2e test for firmware path traversal
A container image could be crafted with a symbolic link in
/lib/firmware/nvidia/ that points to a location outside of the
container's root filesystem. When running such a container with the
NVIDIA Container Toolkit, this could potentially lead to files being
created on the host filesystem.

This change adds an end-to-end test to ensure that the toolkit is not
vulnerable to this kind of path traversal attack.

Signed-off-by: Carlos Eduardo Arango Gutierrez <eduardoa@nvidia.com>