Remove redundant CDI annotations
When generating CDI specs for coherent and noncoherent devices,
we include CDI device annotations. This increases the minimum CDI spec
version to v0.6.0 which is not as widely supported as v0.5.0 or v0.3.0.

This change removes these annotations after splitting the generated
specs to produce coherent and/or non-coherent specs.

Signed-off-by: Evan Lezar <elezar@nvidia.com>