Don't inject enable-cuda-compat hook in CSV mode
When running the nvidia runtime in csv mode, we should not
inject the enable-cuda-compat hook until we make the host driver
version detection more robust on these platforms. As it is, the
host driver version detection returns a version of 1.1 (the suffix
of libcuda.so) meaning that CUDA forward compatibility will always
be used in a container regardless of whether this is required.

This will be properly added to the generated CDI spec for CSV-based
platforms as a follow-up.

Signed-off-by: Evan Lezar <elezar@nvidia.com>