Add support for drop-in config files in a container
This change adds support for drop-in files in a container. Here
the user needs to set RUNTIME_DROP_IN_CONFIG_HOST_PATH in addition
to RUNTIME_DROP_IN_CONFIG. In this case when updating the top-level
containerd config, the imports will use the host path instead of the
container path.

Signed-off-by: Evan Lezar <elezar@nvidia.com>