Disable generation of coherent CDI specs by default
This change disables the functionality for splitting
generated CDI specifications based on device coherence by default.
This was added in #1247, but due to discussions around whether
coherence is an property that should be exposed, we are disabling
this by default.

Note that users can opt in to the feature by running the
`nvidia-ctk cdi generate` with the `--feature-flag=enable-coherent-annotations`
command line flag. Alternatively the `nvidia-ctk cdi generate` command can
be run with the `NVIDIA_CTK_CDI_GENERATE_FEATURE_FLAGS` enviroment set to
include "enable-coherent-annotations" (in a comma-separated list).

Signed-off-by: Evan Lezar <elezar@nvidia.com>