Always update the ldcache in the container
Even though a particular image (distroless, busybox, alpine) doesn't
include ldconfig, a user may include tooling to read the ldcache in the
container or applications may process the ldconfig using dlopen.

This change always creates the /etc/ld.cache file in the container to
ensure that this works as expected.

Signed-off-by: Evan Lezar <elezar@nvidia.com>