Allow file for config source to be specified explicitly
This change allows a "file" config source to also specify the source
path of the file to process. This means that when configuring containerd
with the option --config-source=file=/foo/bar.toml the file /foo/bar.toml
will be used as the source config instead of the output config file.

Signed-off-by: Evan Lezar <elezar@nvidia.com>