Fix containerd drop-in config path
This changes the default drop-in config file path for containerd
from: /etc/containerd/config.d/99-nvidia.toml
to:   /etc/containerd/conf.d/99-nvidia.toml

The default containerd drop-in config path for containerd does not
align with the upstream value. (Which was added after these changes
were made to the toolkit).

Signed-off-by: Evan Lezar <elezar@nvidia.com>