Fix generate tests when cdi hook exists in path
The TestGenerateSpec unit tests resolve the path to the
nvidia-cdi-hook binary if it exists on the system making the
results dependent on system configuration.

This change sets the path to the binary explicitly to skip the
resolution in the path.

Signed-off-by: Evan Lezar <elezar@nvidia.com>
(cherry picked from commit 95a152a6e74d8b0f7ca1536aec5490cf5e522b6c)