Correct typo in nvsandboxutils feature flag
This original implementation of the FeatureDisableNvsandboxUtils feature
flag included a dash that was not supposed to be there. This change
updates the feature flag's string representation to disable-nvsandboxutils.

Special handling is included for users that may still use the old string
value (e.g. for the nvidia-ctk cdi generate command), but no changes are
expected for users of the nvcdi API.

Signed-off-by: Evan Lezar <elezar@nvidia.com>
(cherry picked from commit a37bab416241a23af7cc53e80c83e2d3dea4c280)