Fix update of ldcache for non-matching host and container distros
This change ensures that updating the ldcache in a container also
includes the system search paths for the container ldconfig.

In most cases, the hook will be executing a host ldconfig that may be
configured widely differently from what the container image expects. The
common case is Debian vs non-Debian. But there are also hosts that
configure ldconfig to search in a glibc prefix (e.g. /usr/lib/glibc). To
avoid all these cases, write the container's expected system search
paths to a drop-in conf file that is likely to be last in lexicographic
order. Entries in the top-level ld.so.conf file may be processed after
this drop-in, but this hook does not modify the top-level file if it
exists.

Signed-off-by: Jean-Francois Roy <jeroy@nvidia.com>
Signed-off-by: Evan Lezar <elezar@nvidia.com>