Use enable-cuda-compat hook when ldcache does not exist
This change ensures that the cuda compat drop-in file is created
in a container even if the ld.cache does not exist.

This allows for the creation of the ldcache in the subsequent invocation
of the update-ldcache hook in container images where the cache does not
pre-exist. Examples of such images are distroless images.

Signed-off-by: Evan Lezar <elezar@nvidia.com>
(cherry picked from commit 5f25d95376c28975d57b14ef72b1d036abe0798b)