[no-relnote] Don't add changelog to generated release
Our changelog generation tooling is not robust enough to
be meaningful when generating a GitHub release. This change
removes the call to the generate-changelog script when
creating a github release. This can be added at some point
when we have a more robust mechansim to generate this.

Signed-off-by: Evan Lezar <elezar@nvidia.com>
(cherry picked from commit 63e04edb630af8044a9d1fb5e718d503347305da)